refactor!: unified pty execution strategy
This removes the multiple different ways in which a recipe can be
executed, like jobstart, terminal, dap, etc.

The new solution provides a unified recipe execution, without for
example qf only really working for `build` but `build` not being able to
be opened in a preview or interacted with, like `term`.